Move toolbox archive from lib/tether to cmd/tether
Migrates online toolbox archive code from lib/tether
to cmd/tether. This ensures that we do not override the
file handlers on the applianceVM, only th containerVM.
Fixes #5933.
