+++ cmd/tether/main_linux.go
@@ -15,25 +15,38 @@
 package main
 
 import (
+	"archive/tar"
+	"context"
+	"io"
+	"net/url"
 	"os"
 	"os/signal"
 	"path"
 	"runtime/debug"
+	"strconv"
 	"strings"
 	"syscall"
 
 	log "github.com/Sirupsen/logrus"
 
+	dar "github.com/docker/docker/pkg/archive"
+
+	"github.com/vmware/govmomi/toolbox/hgfs"
+	"github.com/vmware/vic/lib/archive"
+	"github.com/vmware/vic/lib/portlayer/storage/vsphere"
 	"github.com/vmware/vic/lib/tether"
 	viclog "github.com/vmware/vic/pkg/log"
 	"github.com/vmware/vic/pkg/trace"
 	"github.com/vmware/vic/pkg/vsphere/extraconfig"
 )
 
-var tthr tether.Tether
+var (
+	tthr                  tether.Tether
+	defaultArchiveHandler = hgfs.NewArchiveHandler().(*hgfs.ArchiveHandler)
+)
 
 func init() {
-	// Initiliaze logger with default TextFormatter
+	// Initialize logger with default TextFormatter
 	log.SetFormatter(viclog.NewTextFormatter())
 
 	// use the same logger for trace and other logging
@@ -95,7 +108,13 @@ func main() {
 	tthr.Register("Attach", sshserver)
 
 	// register the toolbox extension
-	tthr.Register("Toolbox", tether.NewToolbox().InContainer())
+	toolbox := tether.NewToolbox().InContainer()
+	cmd := toolbox.Service.Command
+	cmd.FileServer.RegisterFileHandler(hgfs.ArchiveScheme, &hgfs.ArchiveHandler{
+		Read:  toolboxOverrideArchiveRead,
+		Write: toolboxOverrideArchiveWrite,
+	})
+	tthr.Register("Toolbox", toolbox)
 
 	err = tthr.Start()
 	if err != nil {
@@ -133,4 +152,117 @@ func startSignalHandler() {
 			}
 		}
 	}()
+}
+
+// toolboxOverrideArchiveRead is the online DataSink Override Handler
+func toolboxOverrideArchiveRead(u *url.URL, tr *tar.Reader) error {
+
+	// special behavior when using disk-labels and filterspec
+	diskLabel := u.Query().Get(vsphere.DiskLabelQueryName)
+	filterSpec := u.Query().Get(vsphere.FilterSpecQueryName)
+	if diskLabel != "" && filterSpec != "" {
+		op := trace.NewOperation(context.Background(), "ToolboxOnlineDataSink: %s", u.String())
+		op.Debugf("Reading from tar archive to path %s: %s", u.Path, u.String())
+		spec, err := archive.DecodeFilterSpec(op, &filterSpec)
+		if err != nil {
+			op.Errorf(err.Error())
+			return err
+		}
+		diskPath, err := MountDiskLabel(diskLabel)
+		if err != nil {
+			op.Errorf(err.Error())
+			return err
+		}
+		defer UnmountDiskLabel(op, diskPath)
+
+		// no need to join on u.Path here. u.Path == spec.Rebase, but
+		// Unpack will rebase tar headers for us. :thumbsup:
+		err = archive.Unpack(op, tr, spec, diskPath)
+		if err != nil {
+			op.Errorf(err.Error())
+		}
+		op.Debugf("Finished reading from tar archive to path %s: %s", u.Path, u.String())
+		return err
+	}
+	return defaultArchiveHandler.Read(u, tr)
+
+}
+
+// toolboxOverrideArchiveWrite is the Online DataSource Override Handler
+func toolboxOverrideArchiveWrite(u *url.URL, tw *tar.Writer) error {
+
+	// special behavior when using disk-labels and filterspec
+	diskLabel := u.Query().Get(vsphere.DiskLabelQueryName)
+	filterSpec := u.Query().Get(vsphere.FilterSpecQueryName)
+
+	skiprecurse, _ := strconv.ParseBool(u.Query().Get(vsphere.SkipRecurseQueryName))
+	skipdata, _ := strconv.ParseBool(u.Query().Get(vsphere.SkipDataQueryName))
+
+	if diskLabel != "" && filterSpec != "" {
+		op := trace.NewOperation(context.Background(), "ToolboxOnlineDataSource: %s", u.String())
+		op.Debugf("Writing to archive from %s: %s", u.Path, u.String())
+
+		spec, err := archive.DecodeFilterSpec(op, &filterSpec)
+		if err != nil {
+			op.Errorf(err.Error())
+			return err
+		}
+
+		// get the container fs mount
+		diskPath, err := MountDiskLabel(diskLabel)
+		if err != nil {
+			op.Errorf(err.Error())
+			return err
+		}
+		defer UnmountDiskLabel(op, diskPath)
+
+		var rc io.ReadCloser
+		if skiprecurse {
+			// we only want a single file - this is a hack while we're abusing Diff, but
+			// accomplish this by generating a single entry ChangeSet
+			changes := []dar.Change{
+				{
+					Kind: dar.ChangeModify,
+					Path: u.Path,
+				},
+			}
+
+			rc, err = archive.Tar(op, diskPath, changes, spec, !skipdata, false)
+		} else {
+			rc, err = archive.Diff(op, diskPath, "", spec, !skipdata, false)
+		}
+
+		if err != nil {
+			op.Errorf(err.Error())
+			return err
+		}
+
+		tr := tar.NewReader(rc)
+		defer rc.Close()
+		for {
+			hdr, err := tr.Next()
+			if err == io.EOF {
+				op.Debugf("Finished writing to archive from %s: %s with error %#v", u.Path, u.String(), err)
+				break
+			}
+			if err != nil {
+				op.Errorf("error writing tar: %s", err.Error())
+				return err
+			}
+			op.Debugf("Writing header: %#s", *hdr)
+			err = tw.WriteHeader(hdr)
+			if err != nil {
+				op.Errorf("error writing tar header: %s", err.Error())
+				return err
+			}
+			_, err = io.Copy(tw, tr)
+			if err != nil {
+				op.Errorf("error writing tar contents: %s", err.Error())
+				return err
+			}
+		}
+
+		return nil
+	}
+	return defaultArchiveHandler.Write(u, tw)
 }
